% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_report.R
\name{anova_report}
\alias{anova_report}
\title{Anova Report}
\usage{
anova_report(data, x, y, browse = TRUE)
}
\arguments{
\item{data}{frame}

\item{x}{grouping variable}

\item{y}{dependent variable}
}
\description{
Example of generating an HTML report from R function
}
\details{
\code{anova_report} generates an HTML report describing a one-way analysis of variance.
The grouping variable is coerced to be a factor.
}
\examples{
data(mpg, package="ggplot2")
anova_report(mpg, class, cty)
anova_report(mpg, class, cty, browse=FALSE)
anova_report(mtcars, cyl, mpg)
}
